Rename AudioFileRepositoryWav to AudioFileRepository_libsndfile to prepare for a refactor for supporting multiple audio file formats in one implementation of the repository
